requirements.txt
streamlit>=1.32.0
google-generativeai>=0.7.0
pandas>=2.2.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
